Revert "SEO for Individual Club Pages"
