+++ frontend/components/DropdownFilter.js
@@ -4,6 +4,7 @@ import {
   CLUBS_GREY_LIGHT,
   CLUBS_BLUE,
   CLUBS_RED,
+  CLUBS_YELLOW,
   BORDER,
   LIGHT_GRAY,
   WHITE,
