+++ frontend/components/common/Modal.js
@@ -93,6 +93,7 @@ export const Modal = ({ show, children, closeModal }) => {
       <Shade className="modal-background" onClick={closeModal} show={show} />
       <ModalCard className="card" onClick={noop} show={show}>
         <CloseModalIcon
+          show={show}
           name="x"
           alt="&#215;"
           onClick={closeModal}
