+++ frontend/components/common/index.js
@@ -4,7 +4,6 @@ export * from './EmptyState'
 export * from './Grid'
 export * from './Icon'
 export * from './Line'
-export * from './Metadata'
 export * from './Modal'
 export * from './Loading'
 export * from './ProfilePic'
