+++ frontend/pages/club/[club]/edit.js
@@ -12,7 +12,6 @@ import {
   formatResponse,
   getRoleDisplay,
 } from '../../../utils'
-import ClubMetadata from '../../../components/ClubMetadata'
 import Form, { ModelForm } from '../../../components/Form'
 import TabView from '../../../components/TabView'
 import AuthPrompt from '../../../components/common/AuthPrompt'
@@ -315,11 +314,7 @@ class ClubForm extends Component {
     const { club, invites, isEdit, message, editMember } = this.state
 
     if (authenticated === false) {
-      return (
-        <AuthPrompt>
-          <ClubMetadata club={club} />
-        </AuthPrompt>
-      )
+      return <AuthPrompt />
     }
 
     if (
@@ -330,7 +325,6 @@ class ClubForm extends Component {
     ) {
       return (
         <AuthPrompt title="Oh no!" hasLogin={false}>
-          <ClubMetadata club={club} />
           You do not have permission to edit the page for {club.name}. To get
           access, contact{' '}
           <a href="mailto:contact@pennclubs.com">contact@pennclubs.com</a>.
@@ -985,7 +979,6 @@ class ClubForm extends Component {
 
     return (
       <Container>
-        <ClubMetadata club={club} />
         <Title>
           {nameOrDefault}
           {showInactiveTag && <InactiveTag />}
