+++ frontend/pages/club/[club]/fair.js
@@ -17,7 +17,6 @@ import {
   Center,
 } from '../../../components/common'
 import Link from 'next/link'
-import ClubMetadata from '../../../components/ClubMetadata'
 
 const Image = s.img`
   height: 86px;
@@ -76,7 +75,6 @@ const Fair = ({
 
   return (
     <PhoneContainer>
-      <ClubMetadata club={club} />
       <Center>
         <ClubHeader>
           {image && <Image src={image} />}
