+++ frontend/pages/club/[club]/flyer.js
@@ -2,15 +2,9 @@ import { useState, useEffect } from 'react'
 import s from 'styled-components'
 
 import { doApiRequest, getApiUrl } from '../../../utils'
-import {
-  Title,
-  Text,
-  Container,
-  Center,
-  Metadata,
-} from '../../../components/common'
+import Head from '../../../components/Header/Head'
+import { Title, Text, Container, Center } from '../../../components/common'
 import { CLUBS_BLUE, CLUBS_NAVY, CLUBS_RED } from '../../../constants/colors'
-import ClubMetadata from '../../../components/ClubMetadata'
 
 const Image = s.img`
   padding: 0;
@@ -144,7 +138,15 @@ const truncate = (str, len = 54) => {
   return `${str.substring(0, len)}...`
 }
 
-const Flyer = ({ query }) => {
+const Flyer = ({
+  authenticated,
+  query,
+  userInfo,
+  favorites,
+  updateFavorites,
+  subscriptions,
+  updateSubscriptions,
+}) => {
   const [clubs, setClubs] = useState(null)
   const [count, setCount] = useState(0)
   const [failedClubs, setFailedClubs] = useState([])
@@ -216,7 +218,7 @@ const Flyer = ({ query }) => {
 
   return (
     <>
-      <Metadata />
+      <Head />
       {!!failedClubs.length && (
         <ErrorPane onClick={() => setShowErrorPane(val => !val)}>
           {showErrorPane ? (
