+++ frontend/pages/invite/[club]/[invite]/[token]/index.js
@@ -1,12 +1,11 @@
-import { Component } from 'react'
-import Link from 'next/link'
-import { withRouter } from 'next/router'
-
 import { CLUB_ROUTE } from '../../../../../constants/routes'
 import renderPage from '../../../../../renderPage'
 import { doApiRequest, formatResponse, LOGIN_URL } from '../../../../../utils'
+import React from 'react'
+import { withRouter } from 'next/router'
+import Link from 'next/link'
 
-class Invite extends Component {
+class Invite extends React.Component {
   constructor(props) {
     super(props)
     this.state = {
