+++ frontend/pages/welcome.js
@@ -7,14 +7,7 @@ import renderPage from '../renderPage'
 import { doApiRequest } from '../utils'
 import ProfileForm from '../components/Settings/ProfileForm'
 import AuthPrompt from '../components/common/AuthPrompt'
-import {
-  Title,
-  Text,
-  Icon,
-  Center,
-  PhoneContainer,
-  Metadata,
-} from '../components/common'
+import { Title, Text, Icon, Center, PhoneContainer } from '../components/common'
 import Link from 'next/link'
 
 const Image = s.img`
@@ -68,69 +61,66 @@ const Welcome = ({ authenticated, query, userInfo, url }) => {
   useEffect(markWelcome, [])
 
   return (
-    <>
-      <Metadata title="Welcome!" />
-      <PhoneContainer>
-        <TitleHeader>
-          <Image src="/static/img/peoplelogo.png" />
-          <Title>Welcome to Penn Clubs!</Title>
-        </TitleHeader>
-        <hr />
-        <Center>
-          <Text>
-            Penn Clubs is your central source of information about student
-            organizations at Penn.
-          </Text>
-        </Center>
-        <hr />
-        <Center>
-          <Subtitle>1. Tell us about yourself</Subtitle>
-          <Text>
-            The info below helps us tailor your Penn Clubs experience to find
-            clubs that you're likely to be interested in. It will also be shared
-            with clubs that you choose to subscribe to. Feel free to leave
-            fields blank if you'd prefer not the share this info.
-          </Text>
-        </Center>
-        <ProfileForm settings={userInfo} />
-        <hr />
-        <Center>
-          <Subtitle>2. Getting started</Subtitle>
-          <Text>
-            Here are two common buttons that you'll see around the site.
-            Bookmarks and subscriptions can be managed from your Penn Clubs
-            account at any time.
-          </Text>
-          <div className="columns is-mobile">
-            <div className="column">
-              <div className="button is-link is-large">
-                <Icon alt="bookmark" name="bookmark" /> Bookmark
-              </div>
-              <Text style={{ marginTop: '0.5rem' }}>
-                To save a club for later
-              </Text>
+    <PhoneContainer>
+      <TitleHeader>
+        <Image src="/static/img/peoplelogo.png" />
+        <Title>Welcome to Penn Clubs!</Title>
+      </TitleHeader>
+      <hr />
+      <Center>
+        <Text>
+          Penn Clubs is your central source of information about student
+          organizations at Penn.
+        </Text>
+      </Center>
+      <hr />
+      <Center>
+        <Subtitle>1. Tell us about yourself</Subtitle>
+        <Text>
+          The info below helps us tailor your Penn Clubs experience to find
+          clubs that you're likely to be interested in. It will also be shared
+          with clubs that you choose to subscribe to. Feel free to leave fields
+          blank if you'd prefer not the share this info.
+        </Text>
+      </Center>
+      <ProfileForm settings={userInfo} />
+      <hr />
+      <Center>
+        <Subtitle>2. Getting started</Subtitle>
+        <Text>
+          Here are two common buttons that you'll see around the site. Bookmarks
+          and subscriptions can be managed from your Penn Clubs account at any
+          time.
+        </Text>
+        <div className="columns is-mobile">
+          <div className="column">
+            <div className="button is-link is-large">
+              <Icon alt="bookmark" name="bookmark" /> Bookmark
             </div>
-            <div className="column">
-              <div className="button is-danger is-large">
-                <Icon alt="subscribe" name="bell" /> Subscribe
-              </div>
-              <Text style={{ marginTop: '0.5rem' }}>
-                To join the mailing list
-              </Text>
-            </div>
+            <Text style={{ marginTop: '0.5rem' }}>
+              To save a club for later
+            </Text>
           </div>
-        </Center>
-        <hr />
-        <Center>
-          <Subtitle>3. Start exploring Penn Clubs!</Subtitle>
-          <Link href={next && next.startsWith('/') ? next : CLUB_ROUTE()}>
-            <a className="button is-success is-large" onClick={markWelcome}>
-              Browse clubs
-            </a>
-          </Link>
-        </Center>
-      </PhoneContainer>
-    </>
+          <div className="column">
+            <div className="button is-danger is-large">
+              <Icon alt="subscribe" name="bell" /> Subscribe
+            </div>
+            <Text style={{ marginTop: '0.5rem' }}>
+              To join the mailing list
+            </Text>
+          </div>
+        </div>
+      </Center>
+      <hr />
+      <Center>
+        <Subtitle>3. Start exploring Penn Clubs!</Subtitle>
+        <Link href={next && next.startsWith('/') ? next : CLUB_ROUTE()}>
+          <a className="button is-success is-large" onClick={markWelcome}>
+            Browse clubs
+          </a>
+        </Link>
+      </Center>
+    </PhoneContainer>
   )
 }
