+++ frontend/stories/Footer.stories.js
@@ -5,3 +5,4 @@ storiesOf('Footer', module)
   .add('Example', () => (
     <Footer />
   ))
+
